VideoCore IV: Avoid an unnecessary COPY in loop instructions
When comparing to an immediate, a copy must be made. Currently, the
value is copied twice: once in loophckcond to check the condition
and once to display the value in the disassembly. However, the
disassembly does not need the value itself; it can use a display
stub and save one unnecessary P-Code.